requirements.txt
streamlit
ultralytics
torch
torchvision
Pillow
pandas
numpy
tqdm
git+https://github.com/openai/CLIP.git